requirements.txt
allure-python-commons
grpcio
grpcio-tools
pytest
pytest-asyncio
allure-pytest